+++ src/atlas/zregionannotationviewsettingtreeview.cpp
@@ -21,7 +21,7 @@ namespace nim {
 ZRegionAnnotationViewSettingTreeView::ZRegionAnnotationViewSettingTreeView(
   ZRegionAnnotationViewSettingTreeModel& objModel,
   ZRegionAnnotation& anno,
-  std::map<int, std::unique_ptr<ZROIFilter>>& idToROIFilters,
+  std::map<int64_t, std::unique_ptr<ZROIFilter>>& idToROIFilters,
   QWidget* parent)
   : ZRegionAnnotationViewSettingTreeView(objModel, anno, parent)
 {
@@ -33,7 +33,7 @@ ZRegionAnnotationViewSettingTreeView::ZRegionAnnotationViewSettingTreeView(
 ZRegionAnnotationViewSettingTreeView::ZRegionAnnotationViewSettingTreeView(
   ZRegionAnnotationViewSettingTreeModel& objModel,
   ZRegionAnnotation& anno,
-  std::map<int, std::unique_ptr<Z3DMeshFilter>>& idToMeshFilters,
+  std::map<int64_t, std::unique_ptr<Z3DMeshFilter>>& idToMeshFilters,
   QWidget* parent)
   : ZRegionAnnotationViewSettingTreeView(objModel, anno, parent)
 {
